+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - Client: Ensure the hostname is used for a DNS request instead of the host:port
 - Client: Mark some errors as 'fatal' such that retrying does not happen
+- Client: Loop using a more fancy restart loop that uses a variable wait time and max restarts
 - Server: Do not spam log by not logging EOF, TCP reader or clean exits
 - General: Rename Proxyguard to ProxyGuard